% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featEng-sd.R
\name{computeRegionSD}
\alias{computeRegionSD}
\title{Computing Region stdev}
\usage{
computeRegionSD(voxel_df, data_df = NULL)
}
\arguments{
\item{voxel_df}{A \code{dataframe} in the following format: the first column
is assumed to be a column labeled "PID," which contains patient
identification. The second, third, and fourth columns are 'x', 'y', and 'z'
columns. The fifth column contains the value of the voxel that corresponds
with the given x-y-z coordinate.}

\item{data_df}{A \code{tibble} that is meant to be outputted from \code{dataDrivenClusters()}.
This tibble should be in the following format: the first three columns are
assumed to be 'x', 'y', and 'z' coordinates. The fourth and fifth columns are
UMAP coordinates, and the sixth column contains cluster labels. If this is
equal to \code{NULL}, \code{computeRegionSD()} will assume that all data comes
from the same cluster.}
}
\value{
A tibble with the following components:
\describe{
\item{pid}{a column indicating patient identification}
\item{sd_c}{columns indicating the calculated stdev of the patient for
the cluster that is specified.}
}
}
\description{
Computes the stdev of the MRI voxel values for each patient
stratified on individual clusters.
}
\details{
\code{computeRegionSD()} takes in two arguments: a \code{tibble}
voxel_df and a \code{tibble} data_df. \code{voxel_df} is assumed to be in long
format, with each individual pixel being represented on a single row, so that
the first columns is 'pid,' columns 2 to 4 are 'x', 'y', 'z', and column 5
is 'value'. \code{data_df} is assumed to be the output of \code{dataDrivenClusters()}.
\code{computeRegionSD()} combines \code{voxel_df} and \code{data_df} into
a single \code{tibble}, and then pivots this combined data with respect to
the columns 'pid' and 'cluster'. We then find the stdev of each pid-cluster
pair. If \code{data_df} is \code{null}, then the function assumes that all
the observations come from the same cluster.
}
