% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featEng-tails.R
\name{computeTailMeans}
\alias{computeTailMeans}
\title{computeTailMeans()}
\usage{
computeTailMeans(voxel_df, data_df = NULL, alpha)
}
\arguments{
\item{voxel_df}{A \code{dataframe} in the following format: the first column
is assumed to be an index column counting from 1 to \code{nrow(voxel_df)},
and the second column is a column labeled "PID," which contains patient
identification. The third, fourth, and fifth columns are 'x', 'y', and 'z'
columns. The sixth column contains the value of the voxel that corresponds
with the given x-y-z coordinate.}

\item{data_df}{A \code{tibble} that is meant to be outputted from \code{dataDrivenClusters()}.
This tibble should be in the following format: the first three columns are
assumed to be 'x', 'y', and 'z' coordinates. The fourth and fifth columns are
UMAP coordinates, and the fifth column contains cluster labels. If this is
equal to \code{null}, \code{computeTailMeans()} will assume that all data comes
from the same cluster.}

\item{@alpha}{A \code{double} that indicates what quantiles should be extracted
from the patient data. The lower the number, the tighter the tail will be.}
}
\description{
Computes the means of the two-sided alpha-percentile tails of
each individual patient stratified on each individual cluster.
}
