% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold-cv.R
\name{crossValidation}
\alias{crossValidation}
\title{Cross Validation}
\usage{
crossValidation(clinical_data, k = 5, n = 5, method = "coxph", id = "PatID")
}
\arguments{
\item{clinical_data}{A long-format \code{data.frame} or \code{tibble} that has
features represented by columns. The function additionally assumes the existence
of a patient ID column, marked by the \code{id} argument in \code{crossValidation()}.}

\item{k}{The number of folds in the k-fold cross validation procedure}

\item{n}{The number of repeats of the k-fold cross validation procedure}

\item{method}{A \code{character} indicating which method will be used in the
cross validation procedure. The default is set to 'coxph', indicating the
cox proportional hazards model. The other available options are ridge and lasso
regression.}

\item{id}{A \code{character} indicating the name of the patient ID column. The
default is assumed to be 'PatID'.}
}
\value{
A list with n elements, one for each repeat. Each element is another
list with the following components:
\describe{
\item{c.indices}{A collection of k concordance indices from that particular repeat}
\item{lowers}{A collection of lower bounds on the confidence intervals for the concordance indices}
\item{uppers}{A collection of upper bounds on the confidence intervals for the concordance indices}
\item{sd}{The standard deviation of c.indices}
\item{mean}{The mean of c.indices}
}
}
\description{
A cross validation procedure with a set amount of repeats towards
the goal of optimizing cox proportional hazards models, lasso models, and ridge
models in the context of features from clinical data.
}
\details{
\code{crossValidation()} runs k-fold cross validation using the specified
model from the argument \code{method} on the clinical data that is provided. In
particular, for each repeat, the function splits the data into k folds. The model
is first fit according to the data in a single fold, and is then tested against the
remaining k-1 folds. The function uses concordance index as a measure of error between
the predicted and actual time-status pairs. We then repeat this process k-1 more times
for each of the other folds, and then repeat this overall process n-1 more times for
each repeat.
}
