% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Incorporating feature engineering with clinical data}
\usage{
mergeData(feat_eng_df, clinical_data, match = "ScrID", id_mapping = NULL)
}
\arguments{
\item{feat_eng_df}{A \code{tibble} that takes the same form as an output of any of
the feature engineering functions in VoxelProc.}

\item{clinical_data}{A \code{dataframe} that includes the clinical data
of the patients. It is expected for the first column to be a patient
identification column.}

\item{match}{A \code{character} indicating which column name corresponds to
the identificaton seen in \code{feat_eng_df} and \code{id_mapping}.}

\item{id_mapping}{A \code{dataframe} that has at least the columns 'match' and
the patient ID seen in \code{clinical_data}, the former of which is an argument
in the function. If 'match' is not specified, then the function will assume that '
ScrID' is the first column name.}
}
\value{
A \code{tibble} that outputs the original clinical dataframe with an
added feature column.
}
\description{
Combines the feature engineering results from \code{computeRegionEntropy()}
with clinical data to form a new data set.
}
\details{
The function takes in three parameters: \code{feat_eng_df} is an output
of \code{computeRegionEntropy()}. \code{clinical_data} is the clinical data
that one wants to merge the entropy data with. \code{id_mapping} is an optional
argument that allows the practitioner to map different forms of patient
identification with each other. If this is \code{null}, then it will be assumed
that the 'pid' column in \code{feat_eng_df} is the same identification as what
is used in \code{clinical_data}. If \code{id_mapping} is not \code{NULL},
then 'pid' will be used as the common identification between \code{feat_eng_df}
and \code{clinical_data}.
}
