% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleClusteredRegions.R
\name{resampleClusteredRegions}
\alias{resampleClusteredRegions}
\title{resampleClusteredRegions}
\usage{
resampleClusteredRegions(
  voxel_df,
  n_pca = 20,
  n_umap = 2,
  n_clust = 2,
  n_resamp = 5,
  subsamp_prop = 0.8,
  region = NULL
)
}
\arguments{
\item{voxel_df}{A \code{list} or \code{matrix} object that represents voxel data.
The first three columns should be coordinate data, while the rest of the
columns represent voxel data.}

\item{n_pca}{A \code{numeric} that indicates how many principal components
will be used in the pipeline. Default is set to 20.}

\item{n_umap}{A \code{numeric} that indicates how many umap components will
be used in the pipeline. Default is set to 2.}

\item{n_clust}{A \code{numeric} that indicates how many clusters the k-means
algorithm will search for in the pipeline. Default is set to 2.}

\item{n_resamp}{A \code{numeric} that indicates how many times
\code{resampleClusteredRegions()} should resample from \code{voxel_df}.}

\item{subsamp_prop}{A \code{numeric} indicating what proportion of \code{voxel_df}
will be resampled at each iteration of \code{resampleClusteredRegions}.}

\item{region}{A \code{character} that indicates which region of the brain the
pipeline will be run on. Default is set to \code{NULL}.}
}
\value{
A list with the following components:
\describe{
\item{Average}{the average of the Rand indices that have been recorded for
each pairwise resampling}
\item{Matrix}{a matrix plot that visualizes the pairwise rand indices of the
resamplings.}
}
}
\description{
Tests the stability of the output of \code{dataDrivenClusters()}
by resampling the voxel data set and using the adjusted Rand Index to compare
cluster values.
}
\details{
\code{resampleClusteringRegions} randomly resamples \code{voxel_df}
according to the arguments \code{n_resamp} and \code{subsamp_prop}. Depending
on these parameters, a certain proportion of the columns of \code{voxel_df}
will be chosen, and \code{dataDrivenClusters()} will be ran on these subsets.
In particular, the resampling will be performed only on columns 4 to \code{ncol(voxel_df)}
of \code{voxel_df}, since the first three columns should be the 'x', 'y',  and 'z'
coordinates. After running \code{dataDrivenClusters} on the subsets, the cluster
labels will be extracted for each resample. \code{resampleClusteringRegions()}
then uses the adjusted Rand Index to find how similar each pair of clusterings are to
each other. The function then averages the pairwise adjusted Rand indices as
a 'score' for how stable the clusterings are. In addition to this a matrix
plot is provided to visualize the pairwise rand indices.
}
