% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featEng-tails.R
\name{computeTailMeans}
\alias{computeTailMeans}
\title{computeTailMeans()}
\usage{
computeTailMeans(voxel_df, alpha = 0.05, data_df = NULL)
}
\arguments{
\item{voxel_df}{A \code{dataframe} in the following format: the first column
is assumed to be an index column counting from 1 to \code{nrow(voxel_df)},
and the second column is a column labeled "PID," which contains patient
identification. The third, fourth, and fifth columns are 'x', 'y', and 'z'
columns. The sixth column contains the value of the voxel that corresponds
with the given x-y-z coordinate.}

\item{alpha}{A \code{double} that indicates what quantiles should be extracted
from the patient data. The lower the number, the tighter the tail will be.}

\item{data_df}{A \code{tibble} that is meant to be outputted from \code{dataDrivenClusters()}.
This tibble should be in the following format: the first three columns are
assumed to be 'x', 'y', and 'z' coordinates. The fourth and fifth columns are
UMAP coordinates, and the sixth column contains cluster labels. If this is
equal to \code{NULL}, \code{computeTailMeans()} will assume that all data comes
from the same cluster.}
}
\value{
A tibble with the following components:
\describe{
\item{pid}{a column indicating patient identification}
\item{cluster}{column indicating which cluster the given feature was extracted
from}
\item{upper_quantile}{the upper alpha percentile of the voxel values in the
given cluster}
\item{lower_quantile}{the lower alpha percentile of the voxel values in the
given cluster}
\item{mean_upper}{the mean of the values above the upper quantile}
\item{mean_lower}{the mean of the values beneath the lower quantile}
}
}
\description{
Computes the means of the two-sided alpha-percentile tails of
each individual patient stratified on each individual cluster.
}
\details{
\code{computeTailMeans()} finds the mean of the alpha-percentile tails
of each individual patient through the following steps: first, the function
extracts the 'PID,', 'x', 'y', 'z', 'value,' and 'cluster' columns from
\code{voxel_df} and \code{data_df}. We run \code{cbind} to combine these specific
columns. Following this, we isolate each individual cluster label and then
extract each individual PID. With each individual patient, we then find the
upper and lower alpha-percentile cutoffs, before finding the means of the
values that lie above and below those cutoffs, respectively. We then collect
these results and put them into a dataframe. If \code{data_df} is not specified,
then the function will assume that all of the data comes from a single cluster.
}
