% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrivenClusters.R
\name{dataDrivenClusters}
\alias{dataDrivenClusters}
\title{Data Driven Clusters}
\usage{
dataDrivenClusters(
  voxel_df,
  n_pca = 20,
  n_umap = 2,
  n_clust = 2,
  region = NULL
)
}
\arguments{
\item{voxel_df}{A \code{list} or \code{matrix} object that represents voxel data.
The first three columns should be coordinate data, while the rest of the
columns represent voxel data.}

\item{n_pca}{A \code{numeric} that indicates how many principal components
will be used in the pipeline. Default is set to 20.}

\item{n_umap}{A \code{numeric} that indicates how many umap components will
be used in the pipeline. Default is set to 2.}

\item{n_clust}{A \code{numeric} that indicates how many clusters the k-means
algorithm will search for in the pipeline. Default is set to 2.}

\item{region}{A \code{character} that indicates which region of the brain the
pipeline will be run on. Default is set to \code{null}.}
}
\value{
A tibble with the following components:
\describe{
\item{data_df}{a tibble that includes the original coordinate values,
the embedded UMAP coordinates, and the cluster value of the voxel data.}
\item{plot}{a scatterplot of the lower-dimensional UMAP embedding, colored
according to the k-means clustering algorithm.}
}
}
\description{
\code{dataDrivenClusters()} applies a pipeline of dimensionality
reduction and cluster analysis to a given set of MRI voxel data. First, we
use principal component analysis to embed the data into some lower-dimensional
space. We then apply UMAP to that embedding and reduce it to yet another
smaller dimension. Lastly, k-means clustering is utilized to separate the
embedded voxel data into separate subgroups. Note that \code{voxel_df}
should be a n x p matrix, where n is the number of coordinates in that
particular region of the brain and p is the number of subjects. The first three
columns of \code{voxel_df} should be the 'x', 'y', and 'z' coordinates,
respectively.
}
