% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrivenClusters.R
\name{dataDrivenClusters}
\alias{dataDrivenClusters}
\title{Data Driven Clusters}
\usage{
dataDrivenClusters(
  voxel_df_long,
  n_pca = 20,
  n_umap = 2,
  n_clust = 2,
  region = NULL
)
}
\arguments{
\item{voxel_df_long}{A long \code{data.frame} or \code{matrix} object that
represents spatial patient-voxel data.}

\item{n_pca}{A \code{numeric} that indicates how many principal components
will be used in the pipeline. Default is set to 20.}

\item{n_umap}{A \code{numeric} that indicates how many umap components will
be used in the pipeline. Default is set to 2.}

\item{n_clust}{A \code{numeric} that indicates how many clusters the k-means
algorithm will search for in the pipeline. Default is set to 2.}

\item{region}{A \code{character} that indicates which region of the brain the
pipeline will be run on. Default is set to \code{NULL}.}
}
\value{
A tibble with the following components:
\describe{
\item{data_df}{a tibble that includes the original coordinate values,
the embedded UMAP coordinates, and the cluster value of the voxel data.}
\item{plot}{a scatterplot of the lower-dimensional UMAP embedding, colored
according to the k-means clustering algorithm.}
}
}
\description{
\code{dataDrivenClusters()} applies a pipeline of dimensionality
reduction and cluster analysis in hopes of isolating signals of interest in
high-dimensional MRI voxel data.
}
\details{
The pipeline uses two layers of dimensionality reduction in
order to isolate signals of interest in high-dimensional voxel data. The first
layer, PCA, will embed the data into a smaller but still non-visualizable
space. The second layer, UMAP, will embed this data into a space specified by
the user. The default for this is the visualizable 2D space. Lastly, k-means
is used to algorithmically determine where the various clusters lie. The
argument \code{voxel_df_long} should be an n by 5 data frame.
The first three columns should be coordinate data, while the rest of the
columns represent voxel data and patient ID values. The column names of the
coordinate values must be \code{x}, \code{y}, and \code{z}, verbatim,
representing the location of the given voxels in the brain.
The column name of the voxel values must be \code{value}, and the column
name of the patient IDs must be \code{pid}. The pipeline runs on the
voxel data, not on the coordinates.
}
