% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Incorporating feature engineering with clinical data}
\usage{
mergeData(feat_eng_df, test_clinical_data, id_mapping = NULL)
}
\arguments{
\item{feat_eng_df}{A \code{tibble} that takes the same form as an output of
\code{computeRegionEntropy()}.}

\item{test_clinical_data}{A \code{dataframe} that includes the clinical data
of the patients.}

\item{id_mapping}{A \code{dataframe} that has columns in the following order:
'SubId', 'PatID', 'ScrID', and 'DataDir'.}
}
\value{
A \code{tibble} that outputs the original clinical dataframe with an
added entropy column.
}
\description{
Combines the feature engineering results from \code{computeRegionEntrop()}
with clinical data to form a new data set.
}
\details{
The function takes in three parameters: \code{feat_eng_df} is an output
of \code{computeRegionEntropy()}. \code{test_clinical_data} is the clinical data
that one wants to merge the entropy data with. \code{id_mapping} is an optional
argument that allows the practitioner to map different forms of patient
identification with each other. If this is \code{null}, then it will be assumed
that the 'pid' column in \code{feat_eng_df} is the same identification as what
is used in \code{test_clinical_data}.
}
